Reduce amount of f32 casts and ensure Draw is adjusted only once

+++ src/ch0_1/show_noise.rs
@@ -18,8 +18,8 @@ struct State {
 impl ExerciseState for State {
     fn new(exercise: &ExerciseData) -> Self {
         State {
-            width: exercise.width(),
-            height: exercise.height(),
+            width: exercise.width() as u32,
+            height: exercise.height() as u32,
             noise: Perlin::new(),
             frames: 0,
         }
